pyproject.toml
[project]
name = "atom-agent"
version = "0.1.0"
description = "Atom: Advanced CLI Coding Agent built on DeepAgents"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4",
    "langgraph>=1.0",
    "langchain>=1.0",
    "langchain-core>=1.0",
    "langchain-anthropic>=1.0",
    "langchain-openai>=1.0",
    "tavily-python>=0.5",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
atom = "atom.cli:main"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"
